requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.1
djangorestframework==3.8.2
idna==2.7
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
urllib3==1.23
